backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
pypdf==5.0.1
google-genai
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
pgvector==0.3.6